add router, components and outlet
