+++ src/components/directory/directory.component.jsx
@@ -9,7 +9,6 @@ const Directory = ({ categories }) => {
       ))}
     </div>
   );
-  
 };
 
 export default Directory;
